+++ chargebee/models/card.py
@@ -12,12 +12,12 @@ class Card(Model):
 
     @staticmethod
     def retrieve(id, env=None):
-        return request.send('get', '/cards/%s' % id, None, env)
+        return request.send('get', request.uri_path("cards",id), None, env)
 
     @staticmethod
     def update_card_for_customer(id, params, env=None):
-        return request.send('post', '/customers/%s/credit_card' % id, params, env)
+        return request.send('post', request.uri_path("customers",id,"credit_card"), params, env)
 
     @staticmethod
     def delete_card_for_customer(id, env=None):
-        return request.send('post', '/customers/%s/delete_card' % id, None, env)
+        return request.send('post', request.uri_path("customers",id,"delete_card"), None, env)