+++ chargebee/models/comment.py
@@ -10,16 +10,16 @@ class Comment(Model):
 
     @staticmethod
     def create(params, env=None):
-        return request.send('post', '/comments', params, env)
+        return request.send('post', request.uri_path("comments"), params, env)
 
     @staticmethod
     def retrieve(id, env=None):
-        return request.send('get', '/comments/%s' % id, None, env)
+        return request.send('get', request.uri_path("comments",id), None, env)
 
     @staticmethod
     def list(params=None, env=None):
-        return request.send('get', '/comments', params, env)
+        return request.send('get', request.uri_path("comments"), params, env)
 
     @staticmethod
     def delete(id, env=None):
-        return request.send('post', '/comments/%s/delete' % id, None, env)
+        return request.send('post', request.uri_path("comments",id,"delete"), None, env)