+++ chargebee/models/customer.py
@@ -13,17 +13,21 @@ class Customer(Model):
 
 
     @staticmethod
+    def create(params=None, env=None):
+        return request.send('post', request.uri_path("customers"), params, env)
+
+    @staticmethod
     def list(params=None, env=None):
-        return request.send('get', '/customers', params, env)
+        return request.send('get', request.uri_path("customers"), params, env)
 
     @staticmethod
     def retrieve(id, env=None):
-        return request.send('get', '/customers/%s' % id, None, env)
+        return request.send('get', request.uri_path("customers",id), None, env)
 
     @staticmethod
     def update(id, params=None, env=None):
-        return request.send('post', '/customers/%s' % id, params, env)
+        return request.send('post', request.uri_path("customers",id), params, env)
 
     @staticmethod
     def update_billing_info(id, params=None, env=None):
-        return request.send('post', '/customers/%s/update_billing_info' % id, params, env)
+        return request.send('post', request.uri_path("customers",id,"update_billing_info"), params, env)