+++ chargebee/models/plan.py
@@ -12,20 +12,20 @@ class Plan(Model):
 
     @staticmethod
     def create(params, env=None):
-        return request.send('post', '/plans', params, env)
+        return request.send('post', request.uri_path("plans"), params, env)
 
     @staticmethod
     def update(id, params=None, env=None):
-        return request.send('post', '/plans/%s' % id, params, env)
+        return request.send('post', request.uri_path("plans",id), params, env)
 
     @staticmethod
     def list(params=None, env=None):
-        return request.send('get', '/plans', params, env)
+        return request.send('get', request.uri_path("plans"), params, env)
 
     @staticmethod
     def retrieve(id, env=None):
-        return request.send('get', '/plans/%s' % id, None, env)
+        return request.send('get', request.uri_path("plans",id), None, env)
 
     @staticmethod
     def delete(id, env=None):
-        return request.send('post', '/plans/%s/delete' % id, None, env)
+        return request.send('post', request.uri_path("plans",id,"delete"), None, env)