+++ chargebee/models/subscription.py
@@ -22,40 +22,40 @@ class Subscription(Model):
 
     @staticmethod
     def create(params, env=None):
-        return request.send('post', '/subscriptions', params, env)
+        return request.send('post', request.uri_path("subscriptions"), params, env)
 
     @staticmethod
     def create_for_customer(id, params, env=None):
-        return request.send('post', '/customers/%s/subscriptions' % id, params, env)
+        return request.send('post', request.uri_path("customers",id,"subscriptions"), params, env)
 
     @staticmethod
     def list(params=None, env=None):
-        return request.send('get', '/subscriptions', params, env)
+        return request.send('get', request.uri_path("subscriptions"), params, env)
 
     @staticmethod
     def subscriptions_for_customer(id, params=None, env=None):
-        return request.send('get', '/customers/%s/subscriptions' % id, params, env)
+        return request.send('get', request.uri_path("customers",id,"subscriptions"), params, env)
 
     @staticmethod
     def retrieve(id, env=None):
-        return request.send('get', '/subscriptions/%s' % id, None, env)
+        return request.send('get', request.uri_path("subscriptions",id), None, env)
 
     @staticmethod
     def update(id, params=None, env=None):
-        return request.send('post', '/subscriptions/%s' % id, params, env)
+        return request.send('post', request.uri_path("subscriptions",id), params, env)
 
     @staticmethod
     def change_term_end(id, params, env=None):
-        return request.send('post', '/subscriptions/%s/change_term_end' % id, params, env)
+        return request.send('post', request.uri_path("subscriptions",id,"change_term_end"), params, env)
 
     @staticmethod
     def cancel(id, params=None, env=None):
-        return request.send('post', '/subscriptions/%s/cancel' % id, params, env)
+        return request.send('post', request.uri_path("subscriptions",id,"cancel"), params, env)
 
     @staticmethod
     def reactivate(id, params=None, env=None):
-        return request.send('post', '/subscriptions/%s/reactivate' % id, params, env)
+        return request.send('post', request.uri_path("subscriptions",id,"reactivate"), params, env)
 
     @staticmethod
     def add_credit(id, params, env=None):
-        return request.send('post', '/subscriptions/%s/add_credit' % id, params, env)
+        return request.send('post', request.uri_path("subscriptions",id,"add_credit"), params, env)