+++ chargebee/request.py
@@ -1,3 +1,4 @@
+import urllib
 from chargebee import util, http
 from chargebee.main import ChargeBee
 
@@ -15,5 +16,8 @@ def send(method, url, params=None, env=None):
     from chargebee.list_result import ListResult
     if 'list' in response:
         return ListResult(response['list'], response.get('next_offset', None))
-
     return Result(response)
+
+def uri_path(*paths):
+    return "/" + "/".join(map(lambda path : urllib.quote_plus(str(path)), paths))
+