+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+### v1.3.0 (2014-06-19)
+* * *
+APIs added:
+* Retrieve invoices for a customer. See https://apidocs.chargebee.com/docs/api/invoices?lang=python#list_invoices_for_a_customer.
+* Retrieve transactions for a customer. See https://apidocs.chargebee.com/docs/api/transactions?lang=python#list_transactions_for_a_customer.
+
+APIs updated:
+* Now, a customer(without subscription) can be charged(Create invoice for Charge) for one time charges. See https://apidocs.chargebee.com/docs/api/invoices?lang=python#create_invoice_for_charge.
+* Now, a customer(without subscription) can be charged for one time addons(Create invoice for Addon). See https://apidocs.chargebee.com/docs/api/invoices?lang=python#create_invoice_for_addon.
+
+### v1.2.9 (2014-05-28)
+* * *
+New API to support Single Sign-on (SSO) to access the customer portal, if you already have your own authentication for your website. See https://apidocs.chargebee.com/docs/api/portal_sessions?lang=python.
+
+### v1.2.8 (2014-05-23)
+* * *
+* New API to create customer without subscription. See https://apidocs.chargebee.com/docs/api/customers#create_a_customer
+
+* New API to fetch invoices for a customer. This helps you fetch the invoices created due to multiple subscriptions present for any customer. See https://apidocs.chargebee.com/docs/api/invoices#list_invoices_for_a_customer
+
+* Customer id reference is added to the invoice attributes.
+
+### v1.2.7  (2014-05-13)
+* * *
+Issue fix: Missing import Download in result.py.
+
 ### v1.2.6  (2014-04-22)
 * * *
 Support for returning shipping address as part of create/update subscription API.